project_actual/requirements.txt
flask==2.1.1
torch==1.10.1
torchvision==0.11.2
clip==1.0
Pillow==9.0.1
numpy==1.22.3
